+++ Sources/Spectre/Global.swift
@@ -1,53 +1,7 @@
 #if os(Linux)
 import Glibc
 #else
-import Darwin
-import XCTest
-
-extension XCTestCase {
-  
-  class Reporter: StandardReporter {
-    
-    weak var testCase: XCTestCase?
-    
-    override func printStatus() {
-      // while running with test case do not print status
-      // only print it in the very end
-      if testCase == nil {
-        super.printStatus()
-      }
-    }
-    
-    override func report(_ name: String, closure: (ContextReporter) -> Void) {
-      if depth == 0 {
-        print("")
-      }
-      super.report(name, closure: closure)
-    }
-    
-    override func addFailure(_ name: String, failure: FailureType) {
-      super.addFailure(name, failure: failure)
-      print("")
-
-      let name = (position + [name]).joined(separator: " ")
-      testCase?.recordFailure(withDescription: "\(name): \(failure.reason)", inFile: failure.file, atLine: failure.line, expected: false)
-    }
-    
-  }
-  
-  static let reporter = Reporter()
-  
-  public func describe(_ name: StaticString = #function, _ test: (ContextType) -> Void) {
-    XCTestCase.reporter.testCase = self
-    defer { XCTestCase.reporter.testCase = nil }
-
-    Spectre.describe(name, closure: test)
-    _ = globalContext.run(reporter: XCTestCase.reporter)
-    globalContext.cases.removeLast()
-  }
-  
-}
-
+  import Darwin
 #endif
 
 let globalContext: GlobalContext = {
@@ -55,8 +9,8 @@ let globalContext: GlobalContext = {
   return GlobalContext()
 }()
 
-public func describe(_ name: StaticString = #function, closure: (ContextType) -> Void) {
-  globalContext.describe(String(describing: name), closure: closure)
+public func describe(_ name: String, closure: (ContextType) -> Void) {
+  globalContext.describe(name, closure: closure)
 }
 
 public func it(_ name: String, closure: @escaping () throws -> Void) {
@@ -71,13 +25,13 @@ public func run() -> Never  {
   } else if CommandLine.arguments.contains("-t") {
     reporter = DotReporter()
   } else {
-    #if os(Linux)
-    reporter = StandardReporter()
+    #if os(tvOS) || os(macOS) || os(iOS)
+      reporter = XcodeReporter()
     #else
-    reporter = XCTestCase.reporter
+      reporter = StandardReporter()
     #endif
   }
-  
+
   run(reporter: reporter)
 }
